feat: add new option `:send-delayed-final-ret`.
`:send-delayed-final-ret': Send the final return with a slight delay.
Some REPLs may not properly recognize when a large chunk of text sent
with bracketed paste mode has finished being input and needs to be
evaluated.  Normally, a final return character signals the REPL to
execute the command, but certain REPLs require a brief delay before
this final return to properly process the bracketed paste input.  This
option should generally remain false (the default), with Claude Code
being a notable exception that requires it to be set to true.
Contributions are welcome if other REPLs are found to need this option
enabled.  The default value is nil.